[KFI]refactor(document-viewer): move comment related components to comment folder
