+++ packages/sn-document-viewer-react/src/components/index.ts
@@ -8,5 +8,5 @@ export { PageList } from './PageList'
 export * from './document-widgets'
 export * from './page-widgets'
 export * from './styles'
-export { CommentComponent } from './Comment'
-export { CreateComment } from './CreateComment'
+export { CommentComponent } from './comment/Comment'
+export { CreateComment } from './comment/CreateComment'
